+++ tests/tests/common/mod.rs
@@ -1,0 +1,4 @@
+
+pub fn common() {
+    println!("this is setup code");
+}
